+++ src/funding/common.jsx
@@ -12,6 +12,7 @@ import {
   COMPONENTS,
   FUNDING,
   ENV,
+  DISPLAY_ONLY_VALUES,
 } from "@paypal/sdk-constants/src";
 import { LOGO_COLOR } from "@paypal/sdk-logos/src";
 
@@ -114,6 +115,7 @@ export type FundingSourceConfig = {|
     layout?: ?$Values<typeof BUTTON_LAYOUT>,
     shippingChange?: ?boolean,
     wallet: ?Wallet,
+    displayOnly?: $ReadOnlyArray<$Values<typeof DISPLAY_ONLY_VALUES>>,
   |}) => boolean,
   Logo: (LogoOptions) => ChildType,
   Mark?: () => ChildType,