+++ src/zoid/buttons/component.jsx
@@ -55,7 +55,6 @@ import {
   isApplePaySupported,
   supportsPopups as userAgentSupportsPopups,
   noop,
-  isLocalStorageEnabled,
 } from "@krakenjs/belter/src";
 import {
   FUNDING,
@@ -80,6 +79,7 @@ import {
   type ButtonProps,
 } from "../../ui/buttons/props";
 import { isFundingEligible } from "../../funding";
+import { CLASS } from "../../constants";
 
 import { containerTemplate } from "./container";
 import { PrerenderedButtons } from "./prerender";
@@ -107,6 +107,7 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
     autoResize: {
       width: false,
       height: true,
+      element: `.${CLASS.AUTORESIZE_CONTAINER}`,
     },
 
     containerTemplate,
@@ -121,18 +122,6 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
 
     prerenderTemplate: ({ state, props, doc, event }) => {
       const { buttonSessionID } = props;
-
-      if (!isLocalStorageEnabled()) {
-        getLogger()
-          .info("localstorage_inaccessible_possible_private_browsing")
-          .track({
-            [FPTI_KEY.BUTTON_SESSION_UID]: buttonSessionID,
-            [FPTI_KEY.CONTEXT_TYPE]: "button_session_id",
-            [FPTI_KEY.CONTEXT_ID]: buttonSessionID,
-            [FPTI_KEY.TRANSITION]:
-              "localstorage_inaccessible_possible_private_browsing",
-          });
-      }
 
       event.on(EVENT.PRERENDERED, () => {
         // CPL stands for Consumer Perceived Latency
@@ -149,17 +138,13 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
             [FPTI_KEY.PAGE]: "main:xo:paypal-components:smart-payment-buttons",
             [FPTI_KEY.CPL_COMP_METRICS]: JSON.stringify(cplPhases?.comp || {}),
           };
-          getLogger()
-            .info("CPL_LATENCY_METRICS_FIRST_RENDER")
-            .track(cplLatencyMetrics);
+          getLogger().track(cplLatencyMetrics);
         } catch (err) {
-          getLogger()
-            .info("button_render_CPL_instrumentation_log_error")
-            .track({
-              err: err.message || "CPL_LOG_PHASE_ERROR",
-              details: err.details,
-              stack: JSON.stringify(err.stack || err),
-            });
+          getLogger().track({
+            err: err.message || "CPL_LOG_PHASE_ERROR",
+            details: err.details,
+            stack: JSON.stringify(err.stack || err),
+          });
         }
       });
 
@@ -202,6 +187,7 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
         createBillingAgreement,
         createSubscription,
         createVaultSetupToken,
+        displayOnly,
       } = props;
 
       const flow = determineFlow({
@@ -244,6 +230,7 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
           supportsPopups,
           supportedNativeBrowser,
           experiment,
+          displayOnly,
         })
       ) {
         return {
@@ -528,6 +515,7 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
             createBillingAgreement,
             createSubscription,
             createVaultSetupToken,
+            displayOnly,
           } = props;
 
           const flow = determineFlow({
@@ -556,6 +544,7 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
               applePaySupport,
               supportsPopups,
               supportedNativeBrowser,
+              displayOnly,
             })
           ) {
             throw new Error(`${fundingSource} is not eligible`);
@@ -829,6 +818,19 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
       onShippingOptionsChange: {
         type: "function",
         required: false,
+      },
+
+      hasShippingCallback: {
+        type: "boolean",
+        required: false,
+        queryParam: true,
+        value: ({ props }) => {
+          return Boolean(
+            props.onShippingChange ||
+              props.onShippingAddressChange ||
+              props.onShippingOptionsChange
+          );
+        },
       },
 
       pageType: {