+++ src/zoid/buttons/util.js
@@ -6,7 +6,6 @@ import {
   isIos,
   isIOS14,
   isSafari,
-  isSFVC,
   type Experiment,
   isDevice,
   isTablet,
@@ -51,7 +50,7 @@ type DetermineFlowOptions = {|
  *
  * @param {string} key for logging
  */
-const logNativeScreenInformation = once((key = "screenInformation") => {
+const logNativeScreenInformation = once(() => {
   if (window) {
     const height = window.innerHeight;
     const outerHeight = window.outerHeight;
@@ -61,18 +60,16 @@ const logNativeScreenInformation = once((key = "screenInformation") => {
     const ios14 = isIOS14();
     const standAlone = isStandAlone();
 
-    const screenInformation = {
-      computedHeight,
-      height,
-      ios14,
-      outerHeight,
-      scale,
-      standAlone,
-    };
-
     getLogger()
       // $FlowFixMe - object is mixed values when this expects all of the same value types
-      .info(key, screenInformation);
+      .info("sfvcScreenInformation", {
+        computedHeight,
+        height,
+        ios14,
+        outerHeight,
+        scale,
+        standAlone,
+      });
   }
 });
 
@@ -91,16 +88,13 @@ export function determineFlow(
 }
 
 export function isSupportedNativeBrowser(): boolean {
+  logNativeScreenInformation();
+
   if (typeof window === "undefined") {
     return false;
   }
 
   if (!userAgentSupportsPopups()) {
-    return false;
-  }
-
-  if (isSFVC()) {
-    logNativeScreenInformation("sfvcScreenInformation");
     return false;
   }
 
@@ -184,6 +178,7 @@ export function getRenderedButtons(
     onShippingChange,
     onShippingAddressChange,
     onShippingOptionsChange,
+    hasShippingCallback,
     style = {},
     enableFunding = getEnableFunding(),
     fundingEligibility = getRefinedFundingEligibility(),
@@ -194,6 +189,7 @@ export function getRenderedButtons(
     createBillingAgreement,
     createSubscription,
     createVaultSetupToken,
+    displayOnly,
   } = props;
 
   const flow = determineFlow({
@@ -217,11 +213,13 @@ export function getRenderedButtons(
     onShippingChange,
     onShippingAddressChange,
     onShippingOptionsChange,
+    hasShippingCallback,
     flow,
     applePaySupport,
     supportsPopups,
     supportedNativeBrowser,
     experiment,
+    displayOnly,
   });
   return renderedButtons;
 }